data-ingestion/requirements.txt
# Google Cloud dependencies
google-cloud-storage>=2.7.0
google-cloud-speech>=2.16.2
google-cloud-translate>=3.10.0
google-cloud-firestore>=2.9.1

# Audio processing
pydub>=0.25.1
librosa>=0.9.2
soundfile>=0.11.0

# Utilities
tqdm>=4.64.1
python-dotenv>=0.21.0
pyyaml>=6.0
click>=8.1.3

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
